+++ src/btreemap/node/v1.rs
@@ -1,0 +1,61 @@
+use super::*;
+
+impl<K: Storable + Ord + Clone> Node<K> {
+    pub(super) fn load_v1<M: Memory>(
+        address: Address,
+        header: NodeHeader,
+        memory: &M,
+        max_key_size: u32,
+        max_value_size: u32,
+    ) -> Self {
+        // Load the entries.
+        let mut keys = Vec::with_capacity(header.num_entries as usize);
+        let mut encoded_values = Vec::with_capacity(header.num_entries as usize);
+        let mut offset = NodeHeader::size();
+        let mut buf = Vec::with_capacity(max_key_size.max(max_value_size) as usize);
+        for i in 0..header.num_entries {
+            // Read the key's size.
+            let key_size = read_u32(memory, address + offset);
+            offset += U32_SIZE;
+
+            // Read the key.
+            buf.resize(key_size as usize, 0);
+            memory.read((address + offset).get(), &mut buf);
+            offset += Bytes::from(max_key_size);
+            let key = K::from_bytes(Cow::Borrowed(&buf));
+            keys.push(key);
+
+            // Values are loaded lazily. Store a reference and skip loading it.
+            encoded_values.push(Value::ByRef(i as u8));
+            offset += U32_SIZE + Bytes::from(max_value_size);
+        }
+
+        // Load children if this is an internal node.
+        let mut children = vec![];
+        if header.node_type == INTERNAL_NODE_TYPE {
+            // The number of children is equal to the number of entries + 1.
+            for _ in 0..header.num_entries + 1 {
+                let child = Address::from(read_u64(memory, address + offset));
+                offset += Address::size();
+                children.push(child);
+            }
+
+            assert_eq!(children.len(), keys.len() + 1);
+        }
+
+        Self {
+            address,
+            keys,
+            encoded_values: RefCell::new(encoded_values),
+            children,
+            node_type: match header.node_type {
+                LEAF_NODE_TYPE => NodeType::Leaf,
+                INTERNAL_NODE_TYPE => NodeType::Internal,
+                other => unreachable!("Unknown node type {}", other),
+            },
+            max_key_size,
+            max_value_size,
+            version: header.version,
+        }
+    }
+}
